chore: moved iterables under utils
