+++ src/chart_types/partition_chart/layout/utils/iterables.ts
